Fixed one more problem with change from Long to Integer for table IDs.  I'm confident that the rest of the problems all relate to the Lucene index, so these are going away at some point in the near future.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@2625 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/tasks/WorkbenchTask.java
@@ -552,11 +552,11 @@ public class WorkbenchTask extends BaseTask
 
     /**
      * Finds the Rollover command that represents the item that has the passed in Record id and
-     * it is checking the CmmandAction's attr by name for the RecordSet
+     * it is checking the CommandAction's attr by name for the RecordSet
      * @param workbench
      * @return
      */
-    protected int getCountOfPanesForTemplate(final Long templateId)
+    protected int getCountOfPanesForTemplate(final Integer templateId)
     {
         int count = 0;
         for (SubPaneIFace sbi : SubPaneMgr.getInstance().getSubPanes())
